refactor: change localStorageKeys filename and change export type for App
